# perf: Add low-level memory ops benchmark