--- a/rs/memory_tracker/BUILD.bazel
+++ b/rs/memory_tracker/BUILD.bazel
@@ -44,3 +44,16 @@ rust_bench(
         "@crate_index//:criterion",
     ] + DEPENDENCIES,
 )
+
+rust_bench(
+    name = "memory_ops_bench",
+    testonly = True,
+    srcs = ["benches/memory_ops.rs"],
+    data = ["benches/test-data/64KiB.txt"],
+    deps = [
+        # Keep sorted.
+        "@crate_index//:criterion",
+        "@crate_index//:memmap2",
+        "@crate_index//:rayon",
+    ] + DEPENDENCIES,
+)