--- a/rs/memory_tracker/Cargo.toml
+++ b/rs/memory_tracker/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "memory_tracker"
-version.workspace = true
 authors.workspace = true
-edition.workspace = true
 description.workspace = true
 documentation.workspace = true
+edition.workspace = true
+name = "memory_tracker"
+version.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -25,9 +25,15 @@ sigsegv_handler_checksum = []
 [dev-dependencies]
 criterion = { workspace = true }
 ic-types = { path = "../types/types" }
+memmap2 = { workspace = true }
 proptest = { workspace = true }
+rayon = { workspace = true }
 tempfile = { workspace = true }
 
 [[bench]]
-name = "traps"
 harness = false
+name = "traps"
+
+[[bench]]
+harness = false
+name = "memory_ops"