… Cargo.lock
… rs/memory_tracker/benches/test-data/64KiB.txt