```
$ bazel run //rs/memory_tracker:memory_ops_bench -- par_iter/threads:8
mmap/par_iter/threads:8/ops:100
                        time:   [1.2654 ms 1.2795 ms 1.2949 ms]
                        thrpt:  [77.227 Kelem/s 78.157 Kelem/s 79.028 Kelem/s]
mmap_mprotect/par_iter/threads:8/ops:100
                        time:   [1.8608 ms 1.8774 ms 1.8937 ms]
                        thrpt:  [52.808 Kelem/s 53.264 Kelem/s 53.740 Kelem/s]
mmap_mprotect_read/par_iter/threads:8/ops:100
                        time:   [4.9503 ms 4.9797 ms 5.0087 ms]
                        thrpt:  [19.965 Kelem/s 20.081 Kelem/s 20.201 Kelem/s]
mmap_mprotect_write/par_iter/threads:8/ops:100
                        time:   [9.0569 ms 9.1043 ms 9.1522 ms]
                        thrpt:  [10.926 Kelem/s 10.984 Kelem/s 11.041 Kelem/s]
mmap_mprotect_read_write/par_iter/threads:8/ops:100
                        time:   [19.574 ms 19.694 ms 19.815 ms]
                        thrpt:  [5.0466 Kelem/s 5.0778 Kelem/s 5.1087 Kelem/s]
mmap_read_write/par_iter/threads:8/ops:100
                        time:   [19.455 ms 19.597 ms 19.733 ms]
                        thrpt:  [5.0676 Kelem/s 5.1028 Kelem/s 5.1400 Kelem/s]
file_read_write/par_iter/threads:8/ops:100
                        time:   [315.32 µs 317.05 µs 318.70 µs]
                        thrpt:  [313.77 Kelem/s 315.41 Kelem/s 317.14 Kelem/s]
```